Add ARM+NEON benchmarking through AWS Graviton
I realized that Scaleway ARM CPUs do not support NEON,
which does not demonstrate the speed abilities of SHISHUA.

I kept it as reference, but simultaneously added support
for benchmarking on Graviton, which does have NEON.

Along the way, it also required adding support for NEON
on ChaCha8, to have a fair benchmark.

I did not add support for xoshiro256+x8,
because its design assume that the compiler itself does
autovectorization. Clang seems to fail to do it right now.
Maybe one day, with better optimizations?

Most benchmarking programs have been catalogued in bin/,
instead of being awkwardly written in the Makefile
with substancial escaping and cumbersome variable tracking.

One particularly thoughtful change in the Makefile
was to support both gcc (which supports disable-tree-cunrolli)
and clang: (which compiles ARM better (thanks Apple and your switch to
M1), but does not support it, and does not understand -march=native).
So we dynamically test whether the $(CC) used is gcc or not.