Enforce longer polling for cloud benchmark setup
The frequent SSH requests will get you blocked client-side otherwise.
The workaround is to kill ssh-agent,
but the longer poll time reduces the chances of this annoyance.

Also, add CPU information in the benchmark output.

Also, improve bash scripts to avoid issues when running from a different
directory.

Also, we take the opportunity to organize bash scripts in bin/