Fix parsing of the seed argument on the CLI
Redditor /u/SJPS noticed an issue on the command-line interface,
which caused random 256-bit seeds provided to SHISHUA from /dev/random
to sometimes produce the same output.

We traced it back to 5bd346c5 (and potentially earlier).
The cause was that,
when the seed argument was longer than 16 characters,
strtoull returned garbage
(plausibly with an error that we did not catch).

We also took the opportunity to absorb up to 256 bits of the provided
seed, instead of just the first 64 bits.